choose_disk: Reworking legacy mode
When using the legacy mode (aka /dev/sd*), the actual code was lacking of
- a good parsing on inputs (should be lists instead of string)
- raw_journal_devices support was missing

The input variables should be lists and not strings as playbook usually defines
as is.

This patch rename the 'legacy_disks' variable to 'devices' which was the current
value. That will make the migration easier.

expand_disk & fake_device functions have a new parameter to indicate the
ceph_type as if it was given by the native syntax. This will allow using the
same code base to filter & report devices.

The output variables should be renamed as we can't override 'devices' and
'raw_journal_devices'.

In the native mode, we'll have :
    - legacy_devices
    - legacy_raw_journal_devices
    - devices_to_activate

In the legacy mode, we'll have :
    - storage_devices
    - journal_devices
    - devices_to_activate

Note that both modes reports the same 'device_to_activate' output variable which
points to already prepared devices which only need activation.