choose_disk: Adding native support for preparing disks
When doing the ceph-disk prepare, we need to point the partition and not
the block device.

In the legacy case, we usually have the name of the disk with a "1"
postfix like "/dev/sdb1".

When using choose_disk, the block device name become a /dev/disk/by-id/
name which have the particular behavior to get partitions named with a
'-part' suffix like : "/dev/disk/by-id/virtio-315e07f6-7ee0-4a82-a"
In that case, the associated parition is "/dev/disk/by-id/virtio-315e07f6-7ee0-4a82-a-part1"

This patch is about splitting the legacy & native associated commands
for activating OSDs.

So in this patch, we have some commands that merge both "legacy_devices" and "storage_devices",
while some does split them in different sub-cases.

Note the native mode offer a simplier code by removing the pretty long
regex_replace for the cciss case.